+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. The format
 is roughly based on [Keep a Changelog], and this project tries to adheres to
 [Semantic Versioning].
 
+## [0.0.29] - TBD
+
+### Changed
+
+- Optional auxilary fields in SAM files `SamTags` is now `SamAuxRaw`.
+- `SamAuxRaw` can now be parsed into `SamAuxData` which supports all optional field types `A`, `i`, `f`, `Z`, `H`, and `B`.
+
 ## [0.0.28] - 2026-04-29
 
 ### Added
@@ -503,6 +510,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - **Added**: Initial internal release. Provides various readers and types for bioinformatics data manipulation.
 
 <!-- Versions -->
+[0.0.29]: https://github.com/CDCgov/zoe/compare/v0.0.28...v0.0.29
 [0.0.28]: https://github.com/CDCgov/zoe/compare/v0.0.27...v0.0.28
 [0.0.27]: https://github.com/CDCgov/zoe/compare/v0.0.26...v0.0.27
 [0.0.26]: https://github.com/CDCgov/zoe/compare/v0.0.25...v0.0.26